+++ query/src/QueryCoordinator.h
@@ -46,7 +46,7 @@ public:
 
     void print_server_configs() const;
 
-    QueryResults send_query_to_workers(const std::string& query);
+    std::pair<QueryResults, size_t> send_query_to_workers(const std::string& query);
 
     static void
     handle_worker_response(const ServerConfig& server_config, QueryResults& results, const std::string& query);
@@ -54,7 +54,8 @@ public:
 private:
     std::vector<ServerConfig> server_configs_;
 
-    static QueryResults handle_worker_response(const ServerConfig& server_config, const std::string& query);
+
+    static std::pair<QueryResults, size_t> handle_worker_response(const ServerConfig& server_config, const std::string& query);
 
     // TODO: Add caching for frequently executed queries
     // TODO: Add query suggestion/autocomplete functionality